Add TransparentHandlerWithOptions function.
Prior to this function being added, it was not possible to set gRPC CallOptions
on the call to `grpc.NewClientStream(...)` performed by the TransparentHandler.
Notably, the `grpc.MaxRecvMsgSize()` would not be configured and would result
in closed streams if the payload size was exceeded, despite the fact that the
proxy's server had the value correctly configured.

The introduction of this `TransparentHandlerWithOptions(...)` function allows
for various settings to be configured on the handler and passed through to the
`grpc.NewClientStream(...)` call.