proxy: properly handle getting an io.EOF from server->client forwards
Without this change, I was able to observe a hang in the proxy in the specific scenario of:

Bidirectional Client -> proxy -> proxy -> Bidirection Server

This change seems to fix this hang that we've seen in practice, though I don't have any sort of self-contained test to
reproduce the hang unfortunately.

I suspect there's an easier reproduction with half-open streams (i.e. using `CloseSend`), though I haven't tried to make
one.